Makes FCS_CKM.1/SKG selection-based.
This is because there very few instances in the NDcPP and the associated
PP modules that actually generate symmetric keys from the RNG directly.
They are usually derived or generated using other mechanisms. There is
one example in the FIA_PSK_EXT.1 that expressly generates a key from the
RNG, and so we justify that usage as being the only applicable selection
that makes sense for the base NDcPP. Other PP-modules might have similar
use cases.

Fundamentally, this SFR seems incorrect -- at least given what the
intent seems to be. It appears to want to say "we use symmetric keys in
this product, so we need to make sure they all come from a trusted
source of randomness." However, in reality, there are very few keys in
most of the SFRs that are derived directly from the RNG as per
SP800-133r2. Instead, we use keys derived from Diffie-Hellman for most
of the protocols. For things like TLS_RSA_WITH_... where the premaster
secret might be transported using an RSA key wrapping, the premaster key
is not actually a symmetric key. It is a CSP that would be used to
derive a series of keys. It's length is 46 random bytes and would be
similar to, say, a PBDKF2 used on a password to ultimately generate an
encryption key used for AES.